+++ python/src/ouster/client/_utils/__init__.py
@@ -1,0 +1,4 @@
+"Python sensor SDK utils"
+# flake8: noqa (unused imports)
+
+from .._client import AutoExposure